+++ CHANGELOG.md
@@ -25,6 +25,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.1] - 2022-02-23
+
+### Added
+- Add a warning log message when no CFS configuration is found that applies to
+  nodes with role Management and subrole Master.
+- Add an info log message that includes the contents of the layer that should be
+  used in the CFS configurations when none are found.
+
 ## [2.0.0] - 2022-02-11
 
 ### Changed